Request_Handler/requirements.txt
# Erstellt mit GitHub Copilot, selbst verfeinert
fastapi>=0.95.0
pydantic>=2.0
openai>=1.0.0
httpx>=0.24.0
python-dotenv>=1.0.0
gunicorn>=20.1.0
uvicorn[standard]>=0.22.0
